+++ src/xyz/oihana/schema/constants/traits/business/documents/InvoiceTrait.php
@@ -11,14 +11,15 @@ namespace xyz\oihana\schema\constants\traits\business\documents;
  */
 trait InvoiceTrait
 {
-    const string ACCOUNT_ID             = 'accountId' ;
-    const string BILLING_PERIOD         = 'billingPeriod' ;
-    const string BROKER                 = 'broker' ;
-    const string CATEGORY               = 'category' ;
-    const string CONFIRMATION_NUMBER    = 'confirmationNumber' ;
-    const string PAYMENT_DUE_DATE       = 'paymentDueDate' ;
-    const string PAYMENT_STATUS         = 'paymentStatus' ;
-    const string PROVIDER               = 'provider' ;
-    const string REFERENCES_ORDER       = 'referencesOrder' ;
-    const string SCHEDULED_PAYMENT_DATE = 'scheduledPaymentDate' ;
+    const string ACCOUNT_ID               = 'accountId' ;
+    const string BILLING_PERIOD           = 'billingPeriod' ;
+    const string BROKER                   = 'broker' ;
+    const string CATEGORY                 = 'category' ;
+    const string CONFIRMATION_NUMBER      = 'confirmationNumber' ;
+    const string PAYMENT_DUE_DATE         = 'paymentDueDate' ;
+    const string PAYMENT_STATUS           = 'paymentStatus' ;
+    const string PROVIDER                 = 'provider' ;
+    const string REFERENCES_DELIVERY_NOTE = 'referencesDeliveryNote' ;
+    const string REFERENCES_ORDER         = 'referencesOrder' ;
+    const string SCHEDULED_PAYMENT_DATE   = 'scheduledPaymentDate' ;
 }